.virtualenv.requirements.txt
#
# Requirement files for Python Virtual Environments
#
# Create a Python virt env with:
#  * For Python2: virtualenv-2.7 .env2
#  * For Python3: virtualenv-3.4 .env3
#
# After activation of the env, install it with:
# $ pip install -U pip setuptools
# $ pip install -r .virtualenv.requirements.txt


docopt
APScheduler
azure-storage
azure-servicemanagement-legacy
python-dateutil

pyOpenSSL
setuptools
future
